+++ .gitignore
@@ -2,6 +2,7 @@
 !tailwind.config.js
 !postcss.config.js
 !jest.config.js
+!esbuild.js
 *.d.ts
 node_modules/
 /build
@@ -36,4 +37,3 @@ coverage
 # Deployment configuration file
 config.yaml
 config-custom.yaml
-/.coverage
