+++ ecs_model_deployer/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "tsc && npm run copy-dist && npm run pack:prod",
+    "build": "node esbuild.js && npm run copy-dist && npm run pack:prod",
     "copy-dist": "cp Dockerfile package.json ./src/cdk*json ../VERSION ./dist",
     "pack:prod": "cd ./dist && npm i --omit dev",
     "clean": "rm -rf ./dist/",
@@ -13,12 +13,13 @@
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
-    "@cdklabs/cdk-enterprise-iac": "^0.0.512",
     "aws-cdk": "^2.177.0",
-    "aws-cdk-lib": "^2.185.0",
+    "aws-cdk-lib": "^2.177.0",
     "zod": "^3.24.1"
   },
   "devDependencies": {
+    "esbuild": "^0.25.1",
+    "@cdklabs/cdk-enterprise-iac": "^0.0.512",
     "@types/node": "^22.4.1"
   }
 }
