+++ ecs_model_deployer/src/cdk.json
@@ -1,5 +1,5 @@
 {
-  "app": "node ./ecs_model_deployer/src/lib/app.js",
+  "app": "node ./lib/index.js",
   "requireApproval": "never",
   "watch": {
     "include": ["**"],
