+++ lib/api-base/fastApiContainer.ts
@@ -70,7 +70,7 @@ export class FastApiContainer extends Construct {
 
         const { config, securityGroup, tokenTable, vpc } = props;
         const buildArgs: Record<string, string> | undefined = {
-            BASE_IMAGE: 'python:3.11',
+            BASE_IMAGE: config.baseImage,
             PYPI_INDEX_URL: config.pypiConfig.indexUrl,
             PYPI_TRUSTED_HOST: config.pypiConfig.trustedHost,
             LITELLM_CONFIG: yamlDump(config.litellmConfig),
@@ -97,6 +97,11 @@ export class FastApiContainer extends Construct {
             environment.TOKEN_TABLE_NAME = tokenTable.tableName;
         }
         const restApiPath = path.join(HERE, '..', 'serve', 'rest-api');
+        const image = config.restApiConfig.imageConfig || {
+            baseImage: config.baseImage,
+            path: restApiPath,
+            type: EcsSourceType.ASSET
+        };
         const apiCluster = new ECSCluster(scope, `${id}-ECSCluster`, {
             config,
             ecsConfig: {
@@ -116,11 +121,7 @@ export class FastApiContainer extends Construct {
                 },
                 buildArgs,
                 containerConfig: {
-                    image: {
-                        baseImage: 'python:3.11',
-                        path: restApiPath,
-                        type: EcsSourceType.ASSET
-                    },
+                    image,
                     healthCheckConfig: {
                         command: ['CMD-SHELL', 'exit 0'],
                         interval: 10,
