+++ lib/core/layers/common/create_env_variables.py
@@ -19,7 +19,11 @@ region_name = os.environ["AWS_REGION"]
 if "iso" in region_name:
     cacerts_directory = "/etc/pki/ca-trust/extracted/pem"
     merged_cert_filename = os.path.join(cacerts_directory, "tls-ca-bundle.pem")
-    os.environ["AWS_CA_BUNDLE"] = merged_cert_filename
-    os.environ["REQUESTS_CA_BUNDLE"] = merged_cert_filename
-    os.environ["SSL_CERT_FILE"] = merged_cert_filename
-    os.environ["SSL_CERT_DIR"] = cacerts_directory
+    if not os.environ.get("AWS_CA_BUNDLE"):
+        os.environ["AWS_CA_BUNDLE"] = merged_cert_filename
+    if not os.environ.get("REQUESTS_CA_BUNDLE"):
+        os.environ["REQUESTS_CA_BUNDLE"] = merged_cert_filename
+    if not os.environ.get("SSL_CERT_DIR"):
+        os.environ["SSL_CERT_DIR"] = cacerts_directory
+    if not os.environ.get("SSL_CERT_FILE"):
+        os.environ["SSL_CERT_FILE"] = merged_cert_filename
