+++ lib/docs/package.json
@@ -14,7 +14,6 @@
   "author": "",
   "license": "Apache-2.0",
   "devDependencies": {
-    "vitepress": "^1.6.3",
-    "vite": "6.2.3"
+    "vitepress": "^1.6.3"
   }
 }
