+++ lib/rag/state_machine/ingest-pipeline.ts
@@ -155,7 +155,7 @@ export class IngestPipelineStateMachine extends Construct {
         }));
 
         policyStatements.map((policyStatement) => ingestPipelineRole.addToPolicy(policyStatement));
-        const lambdaPath = path.join(HERE, '..', '..','..', 'lambda');
+        const lambdaPath = config.lambdaPath || path.join(HERE, '..', '..','..', 'lambda');
 
         // Function to list objects modified in last 24 hours
         const listModifiedObjectsFunction = new Function(this, 'listModifiedObjectsFunc', {
