+++ lib/user-interface/react/vite.config.ts
@@ -21,7 +21,7 @@ import { resolve } from 'node:path';
 // https://vitejs.dev/config/
 export default defineConfig({
     plugins: [react(), splitVendorChunkPlugin()],
-    base: '/prod',
+    base: '/prod/',
     server: {
         port: 3000,
     },
