+++ package.json
@@ -59,6 +59,7 @@
     "@typescript-eslint/parser": "^6.21.0",
     "aws-cdk": "^2.184.0",
     "depcheck": "^1.4.7",
+    "esbuild": "^0.25.1",
     "eslint": "^8.57.1",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-react-hooks": "^4.6.2",
