+++ test/cdk/stacks/nag.test.ts
@@ -33,7 +33,7 @@ enum NagType {
 const nagResults: NagResult = {
     LisaApiBase: [1,7,0,5],
     LisaApiDeployment: [0,0,0,0],
-    LisaChat: [2,33,0,13],
+    LisaChat: [2,20,0,18],
     LisaCore: [0,0,0,0],
     LisaDocs: [1,23,0,13],
     LisaIAM: [0,14,0,0],
